+++ app/models/__init__.py
@@ -1,0 +1,8 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+
+def configure(app):
+    db.init_app(app)
+    app.db = db
