+++ app/schemas/__init__.py
@@ -1,0 +1,7 @@
+from flask_marshmallow import Marshmallow
+
+marshmallow = Marshmallow()
+
+
+def configure(app):
+    marshmallow.init_app(app)
