+++ app/schemas/product.py
@@ -1,0 +1,7 @@
+from ..models.product import Product
+from ..schemas import marshmallow
+
+
+class ProductSchema(marshmallow.SQLAlchemyAutoSchema):
+    class Meta:
+        model = Product
